services/backend/requirements.txt
pandas
numpy
Flask
gunicorn
flask_restx
python-dotenv
requests
beautifulsoup4
datasets
transformers==4.48.0
torch
faiss-cpu
prefect
sentence-transformers
dvc-gs
langchain-text-splitters
langchain-community
google-cloud-storage
mlflow
langchain[mistralai]
langgraph
langfair
datetime
flask-cors
langchain-huggingface
